finish assignment 17

+++ src/editor/view.rs
@@ -267,8 +267,12 @@ impl View {
         self.needs_redraw = true;
     }
     fn backspace(&mut self) {
-        self.move_left();
-        self.delete();
+        if self.text_location.line_index != 0 || self.text_location.grapheme_index != 0
+        {
+            self.move_text_location(&Direction::Left);
+            self.delete();
+        }
+
     }
     fn delete(&mut self) {
         self.buffer.delete(self.text_location);
